Refactors, and making sure 3rd hover works
